outsourced search api methods
