+++ backend/api/api.py
@@ -1,25 +1,9 @@
 from bottle import route, run, response
 import bottle
-from pymongo import MongoClient
-import pymongo
-from bson.objectid import ObjectId
-import json
-from json import JSONEncoder
-try:
-    reduce
-except:
-    from functools import reduce
-from search import *
 
+# import lib files
+from lib import search
 
-#JSON Encoder
-# https://stackoverflow.com/questions/28251835/from-pymongo-objectid-import-objectid-importerror-no-module-named-objectid
-class MongoEncoder(JSONEncoder):
-    def default(self, obj, **kwargs):
-        if isinstance(obj, ObjectId):
-            return str(obj)
-        else:            
-            return JSONEncoder.default(obj, **kwargs)
 
 # source: https://stackoverflow.com/questions/17262170/bottle-py-enabling-cors-for-jquery-ajax-requests
 # the decorator
@@ -41,28 +25,8 @@ def enable_cors(fn):
 def find(query):
     response.headers['Content-type'] = 'application/json'
 
-    client = MongoClient('mongodb://localhost:27017/')
-    db = client["LanguageLevelSearchEngine"]
-    col = db["news_de_DE"]
-
-    terms = query.split()
-    docs = getIdsFromWord(terms)
-    docIds = [id[0] for id in docs]
-
-    documents = []
-    for id in docIds:
-        query= {"_id": ObjectId(id)}
-        documents.append(col.find(query)[0])
-
-    return json.dumps(documents, cls=MongoEncoder)
+    return search.findDocuments(query)
 
 
-def getIdsFromWord(terms):
-    client = MongoClient('mongodb://localhost:27017/')
-    db = client["LanguageLevelSearchEngine"]
-    # TODO: Allow multiple terms search
-    objdb = db["inverted_index_de_DE"].find({'word': terms[0]}).skip(0)
-    entries = [entry for entry in objdb]
-    return entries[0]["documents"]
 
 run(host='localhost', port=8080, debug=True)
